Many things news
Now stamina have a limit between 0 to 100.

Trying to make the BP_Characters follow the one the player is controlling. Wanna create a Controll in with the NPCS Stop following you when you give a command and get close to you when you press again.

Trying to modify the heavy. This will make the heavy attack make more damage than the quick attack.